authorization

+++ backend/db/repository/blog.py
@@ -26,7 +26,9 @@ def list_all_active_blogs(db: Session):
 def update_blog_by_id(id: int, blog: UpdateBlog, db: Session, author_id: int = 1):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {"error": f"Blog with id {id} does not exist"}
+    if not blog_in_db.author_id == author_id:
+        return {"error": f"Only the author can modify a blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
@@ -35,9 +37,12 @@ def update_blog_by_id(id: int, blog: UpdateBlog, db: Session, author_id: int = 1
 
 
 def delete_blog_by_id(id: int, db: Session, author_id: int):
+    print("--> Delete a Blog")
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db.first():
         return {"error": f"Couldn't find blog with id {id}"}
+    if not blog_in_db.first().author_id == author_id:
+        return {"error": f"Only the author can delete a blog"}
     blog_in_db.delete()
     db.commit()
     return {"message": f"Deleted blog with id {id}"}
